refactor: 모달을 공통 컴포넌트로 이동
